third_party/blink/web_tests/fast/forms/calendar-picker/date-picker-with-enter-key.html

<!DOCTYPE html>
<meta name=fuzzy content="maxDifference=0-3; totalPixels=0-1000">
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>


<input type=date id=date value="2019-09-10">

<script>

test(() => {
  document.getElementById('date').focus();
  eventSender.keyDown("Enter");
  assert_equals(internals.pagePopupWindow, null);
}, "The date picker should not open with the Enter key.");

</script>
